### 2017-03-08 #103 @mpadge
osmdata package
Download OpenStreetMap (OSM) data and convert to native R objects defined by sf or sp packages. OSM datasets are extracted from the overpass API and processed with efficient C++ routines.

```
Package: osmdata
Version: 0.0.0
    person("Mark", "Padgham", email="mark.padgham@email.com", role=c("aut", "cre")),
    person("Bob", "Rudis", role="aut"),
    person("Robin", "Lovelace", role="aut"),
    person("Maëlle", "Salmon", role="aut"),
    person("Andrew", "Smith", role="ctb"),
    person("Marcin", "Kalicinski", role=c("ctb", "cph"),
        comment="Author of included RapidXML code"),
    person("Finkelstein", "Noam", role=c("ctb","cph"),
        comment="Author of included stub.R code"),
    person ("Bartnik", "Lukasz", role=c("ctb","cph"),
        comment="Author of included stub.R code"))
Maintainer: Mark Padgham <mark.padgham@email.com>
Title: Download OpenStreetMap Data with the Overpass API
Description: Extraction of OpenStreetMap (OSM) data as 'sf' or 'sp' objects.
    OSM data are extracted from the overpass API and processed with very fast
    C++ routines for return to R.
License: GPL-3 + file LICENSE
SystemRequirements: C++11
NeedsCompilation: yes
    curl,
    httr,
    Rcpp (>= 0.12.4),
    rvest,
    sp,
    xml2
    devtools,
    microbenchmark,
    roxygen2,
    sf,
LinkingTo: Rcpp
```

 Anyone interested in geospatial analysis and visualisation, especially geographers, urban and transport modellers and planners.

The package is similar to `GDAL` through the interface provided by `sf`, yet produces fundamentally different results, as briefly described in the [main vignette](https://osmdatar.github.io/osmdata/articles/osmdata.html), and detailed at length in a [second vignette](https://osmdatar.github.io/osmdata/articles/osm-sf-translation.html). `osmdata` also reproduces much functionality available in the CRAN package `osmar`, yet does so enormously faster (like at least thousands of times faster!) and with a more intuitive user interface. The package also imports data faster than `sf`. `osmar` only returns data defined by a bounding box, while the `osmdata` interface to the overpass API allows highly customised and specific queries to be constructed (see examples in [main vignette](https://osmdatar.github.io/osmdata/articles/osmdata.html)). `osmdata` is also designed to complement the `ropensci` package [`osmplotr`](https://github.com/ropensci/osmplotr), and the latter package will be modified immediately following CRAN acceptance of `osmdata`.

NOTE: As per the JOSS guidelines, the package will be submitted to a long-term repo subsequent to the review process

   - Jochen Topf / joto - Main developer of [osmcode](https://github.com/osmcode) and author of [libosmium](https://github.com/osmcode/libosmium)
   - Edzer Pebesma / edzer - Developer of the `R` packages `sp` and `sf`
