### 2017-03-09 #103 @mpadge
One further comment prior to review: The `goodpractice` output about `<-` instead of `=` and also about `sapply()` mostly arise from fragments of `sf` code which are copied verbatim to ensure absolute consistency with `sf` functionality. I do not intend to modify these fragments (except to reflect changes in `sf` itself). Other than this, `sapply()`is only used in `sapply (x, is.null)`-type statements, to which the `goodpractice` remarks do not apply.
