### 2017-03-09 cbe5218
fix DUMP_INPUT for osmdata-sf to osmdata_sf.xml
