### 2017-03-09 d640688
Suggest vertical alignment of braces.
