+++ R/zzz.R
@@ -1,3 +1,5 @@
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Data (c) OpenStreetMap contributors, ODbL 1.0. http://www.openstreetmap.org/copyright")
+    msg <- paste0 ("Data (c) OpenStreetMap contributors,",
+                   " ODbL 1.0. http://www.openstreetmap.org/copyright")
+    packageStartupMessage (msg)
 }
