+++ vignettes/osmdata.Rmd
@@ -6,8 +6,10 @@ author:
 date: "`r Sys.Date()`"
 bibliography: osmdata-refs.bib
 output: 
-  rmarkdown::html_vignette:
-    toc: true
+    html_document:
+        toc: true
+        toc_float: true
+        theme: flatly
 vignette: >
   %\VignetteIndexEntry{1. osmdata}
   %\VignetteEngine{knitr::rmarkdown}
