### 2017-03-12 738c7a9
more informative error msg for c (sp)
