### 2017-03-12 cea6c4b
add testthat/test-c.R
