### 2017-03-14 6dde21b
fix spelling mistake in man/overpass_status
