### 2017-03-14 19b76d3
update get-osmdata.R (again)
