+++ R/osmdata-package.R
@@ -53,7 +53,7 @@
 #' @importFrom methods is
 #' @importFrom Rcpp evalCpp
 #' @importFrom rvest html_attr html_nodes
-#' @importFrom utils read.table timestamp
+#' @importFrom utils read.table 
 #' @importFrom xml2 read_html read_xml xml_attr xml_text xml_find_all
 #' @useDynLib osmdata
 #' @importFrom jsonlite fromJSON
