+++ NAMESPACE
@@ -37,7 +37,6 @@ importFrom(methods,is)
 importFrom(rvest,html_attr)
 importFrom(rvest,html_nodes)
 importFrom(utils,read.table)
-importFrom(utils,timestamp)
 importFrom(xml2,read_html)
 importFrom(xml2,read_xml)
 importFrom(xml2,xml_attr)
