### 2017-03-20 d7931fd
add @noRd to roxygen comments
