### 2017-03-23 #103 @karthik
### Editor checks:

---

#### Editor comments

Thank you for your submission! Out of curiosity, can you speak to how this relates to functionality from the [Openstreetmap](https://cran.r-project.org/web/packages/OpenStreetMap/index.html) package in terms of overlap?

While I seek reviewers I want to bring your attention to a few checks that were reuturned by `goodpractice`. I have noted your comments about the notes around code copied over from sf (so I've removed those). There is no urgency to address them before the review, but it would be good to fix during.

-  Here are outputs from automated testing using `goodpractice::gp()`. I'd also recommend using `devtools::spell_check()` to examine documentation.

```
── GP osmdata ───────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all
    package code in general. 84% of code lines are
    covered by test cases.

    R/features.R:24:NA
    R/features.R:55:NA
    R/opq.R:75:NA
    R/opq.R:76:NA
    R/osm-extract.R:55:NA
    ... and 275 more lines

  ✖ avoid long code lines, it is bad for readability.
    Also, many people prefer editor windows that are
    about 80 characters wide. Try make your lines
    shorter than 80 characters

    tests/testthat/test-extract.R:74:1
    tests/testthat/test-extract.R:84:1
    tests/testthat/test-extract.R:121:1
    tests/testthat/test-extract.R:171:1
    tests/testthat/test-osmdata.R:21:1
    ... and 15 more lines

  ✖ not import packages as a whole, as this can cause
    name clashes between the imported packages.
    Instead, import only the specific functions you
    need.
```
