### 2017-03-23 a01a583
fix getbb man entry
