### 2017-03-23 #103 @mpadge
Thanks Karthik!

In response to your Q, `OpenStreetMap` simply produces raster layers representing OSM data, and is a purely graphical thing; `osmdata` returns the actual data, and is not a graphical thing at all. I see no real relationship between them.

As for the goodpractice results:
* This is a package which can not be fully covered in tests, but is actually [> 90% in total](https://codecov.io/gh/osmdatar/osmdata), and will increase *slightly* beyond that
* The import warning is for `sp`, and can't be avoided because the `/src` code needs an `[[Rcpp::Depends(sp)]]`, requiring the whole package to be imported even though it's not actually required in the `R` code at all.
* Long lines are only for tests, and are all for long strings containing warning/error messages being tested. Can be eliminated by reducing indentation within each test block, but believe me, that makes the tests much harder to read, not easier. They are also almost all only 81 or 82 characters. I would strongly prefer not to change these, if that's acceptable.
