### 2017-03-24 #103 @noamross
FYI, while we don't strictly enforce line limits, putting `#nolint` after a long message line or URL in code suppresses linting that line and gives me peace of mind and clean `gp()` results.  A [`.lintr` file](https://github.com/ecohealthalliance/fasterize/blob/master/.lintr) is useful for excluding files in some cases (such as auto-generated Rcpp code), but of course easily abused ;)
