### 2017-03-24 #103 @karthik
@mpadge  Perfect, your responses make sense to me.
@noamross Thanks for those useful tips! 💯
