### 2017-03-24 #103 @mpadge
Thanks @noamross, i didn't know about `# nolint` and will definitely modify accordingly. I had [`syntastic`](https://github.com/vim-syntastic/syntastic)ed all source files, but forgot about the `tests`, which actually have lots more tidying issues than just the `gp` ones ... whoops

Thanks @karthik for arranging one reviewer so quickly, and thanks @edzer for agreeing to review.
