### 2017-03-26 07f76c3
fix syntastic checks for all tests
