+++ tests/testthat/test-getbb.R
@@ -8,23 +8,24 @@ get_local <- FALSE
 if (get_local) {
     # vcr code, for when it eventually appears on CRAN:
     #bb_test <- getbb ("Salzburg")
-    #saveRDS (bb_test, 
+    #saveRDS (bb_test,
     #         file = "./tests/testthat/bb_test.Rds")
 
     # Equivalent code using internal 'stub.R' function
-    base_url = "https://nominatim.openstreetmap.org"
+    base_url <- "https://nominatim.openstreetmap.org"
     query <- list (q = 'Salzburg', viewbox = NULL, format = 'json',
                    featuretype = 'settlement', key = NULL, limit = 10)
-    the_url <- httr::modify_url (base_url, query=query)
+    the_url <- httr::modify_url (base_url, query = query)
     cfm_output_bb <- NULL
     trace(
           curl::curl_fetch_memory,
-          exit = function() { cfm_output_bb <<- returnValue() }
-          )
+          exit = function() {
+              cfm_output_bb <<- returnValue()
+          })
     res <- httr::GET (the_url)
     class (cfm_output_bb) <- 'response'
     untrace (curl::curl_fetch_memory)
-    save (cfm_output_bb, file='../cfm_output_bb.rda')
+    save (cfm_output_bb, file = '../cfm_output_bb.rda')
 }
 
 context ("bbox")
@@ -33,7 +34,7 @@ test_that ("bbox", {
   expect_error (bbox_to_string (), "bbox must be provided")
   #expect_error (bbox_to_string ("a"), "bbox must be numeric")
   expect_error (bbox_to_string (1:3), "bbox must contain four elements")
-  expect_message (bbox_to_string (1:5), 
+  expect_message (bbox_to_string (1:5),
                               "only the first four elements of bbox used")
 })
 
@@ -43,11 +44,11 @@ test_that ('getbb-place_name', {
                    {
                        load("../cfm_output_bb.rda")
                        stub (getbb, 'httr::GET', function (x) cfm_output_bb )
-                   } 
-                   res <- getbb (place_name="Salzburg")
+                   }
+                   res <- getbb (place_name = "Salzburg")
                    expect_is (res, "matrix")
                    expect_length (res, 4)
-                   res <- getbb (place_name="Salzburg", format_out="string")
+                   res <- getbb (place_name = "Salzburg", format_out = "string")
                    expect_is (res, "character")
                }
           })
