+++ tests/testthat/test-sf-construction.R
@@ -9,8 +9,8 @@ make_sfc <- function (x, type) {
         xy <- do.call (rbind, do.call ("c", x))
     else
         xy <- do.call (rbind, x)
-    xvals <- xy [,1]
-    yvals <- xy [,2]
+    xvals <- xy [, 1]
+    yvals <- xy [, 2]
     bb <- structure(rep(NA_real_, 4), names = c("xmin", "ymin", "xmax", "ymax"))
     bb [1:4] <- c (min (xvals), min (yvals), max (xvals), max (yvals))
     if (type == "POLYGON")
@@ -18,15 +18,17 @@ make_sfc <- function (x, type) {
     else if (grepl ("MULTI", type) & !is.list (x [[1]]))
         x <- list (x)
     if (type != "MULTIPOLYGON")
-        x <- lapply (x, function (i) structure (i, class = c ("XY", type, "sfg")))
+        x <- lapply (x, function (i)
+                     structure (i, class = c ("XY", type, "sfg")))
     else
-        x <- lapply (x, function (i) structure (list (i), 
-                                                class = c ("XY", type, "sfg")))
+        x <- lapply (x, function (i)
+                     structure (list (i), class = c ("XY", type, "sfg")))
     attr (x, "n_empty") <- sum(sapply(x, function(x) length(x) == 0))
     class(x) <- c(paste0("sfc_", class(x[[1L]])[2L]), "sfc")
-    attr(x, "precision") = 0.0
+    attr(x, "precision") <- 0.0
     attr(x, "bbox") <- bb
-    NA_crs_ <- structure(list(epsg = NA_integer_, proj4string = NA_character_), class = "crs")
+    NA_crs_ <- structure(list(epsg = NA_integer_,
+                              proj4string = NA_character_), class = "crs")
     attr(x, "crs") <- NA_crs_
     x
 }
@@ -34,7 +36,7 @@ make_sfc <- function (x, type) {
 make_sf <- function (...)
 {
     x <- list (...)
-    sf = sapply(x, function(i) inherits(i, "sfc"))
+    sf <- sapply(x, function(i) inherits(i, "sfc"))
     sf_column <- which (sf)
     if (!is.null (names (x [[sf_column]])))
         row.names <- names (x [[sf_column]])
@@ -43,18 +45,18 @@ make_sf <- function (...)
     df <- if (length(x) == 1) # ONLY sfc
                 data.frame(row.names = row.names)
             else # create a data.frame from list:
-                    data.frame(x[-sf_column], row.names = row.names, 
+                    data.frame(x[-sf_column], row.names = row.names,
                            stringsAsFactors = TRUE)
 
-    object <- as.list(substitute(list(...)))[-1L] 
+    object <- as.list(substitute(list(...)))[-1L]
     arg_nm <- sapply(object, function(x) deparse(x))
     sfc_name <- make.names(arg_nm[sf_column])
     #sfc_name <- "geometry"
     df [[sfc_name]] <- x [[sf_column]]
     attr(df, "sf_column") <- sfc_name
-    f = factor(rep(NA_character_, length.out = ncol(df) - 1), 
+    f <- factor(rep(NA_character_, length.out = ncol(df) - 1),
                levels = c ("constant", "aggregate", "identity"))
-    names(f) = names(df)[-ncol (df)]
+    names(f) <- names(df)[-ncol (df)]
     attr(df, "agr") <- f
     class(df) <- c("sf", class(df))
     return (df)
@@ -65,13 +67,13 @@ make_sf <- function (...)
 # **********************************************************
 
 test_that ("sfg-point", {
-               x <- structure (1:2, class=c("XY", "POINT", "sfg"))
[…]

[…]
