### 2017-03-26 ccfe768
sapply->vapply gp()-conform
