+++ tests/timing-benchmark.R
@@ -1,11 +1,11 @@
 #library (osmdata)
-# NOTE: Currently requires devtools::load_all ("osmdata", export_all=TRUE)!
+# NOTE: Currently requires devtools::load_all ("osmdata", export_all = TRUE)!
 
-benchmark <- function (times=100)
+benchmark <- function (times = 100)
 {
-    devtools::load_all (".", export_all=FALSE)
-    q0 <- opq (bbox=c(-0.27,51.47,-0.20,51.50)) 
-    q1 <- add_feature (q0, key='name', value="Thames", exact=FALSE)
+    devtools::load_all (".", export_all = FALSE)
+    q0 <- opq (bbox = c(-0.27, 51.47, -0.20, 51.50))
+    q1 <- add_feature (q0, key = 'name', value = "Thames", exact = FALSE)
     # contains both multipolygons and multilinestrings
     doc <- osmdata_xml (q1, "export.osm")
 
@@ -15,8 +15,8 @@ benchmark <- function (times=100)
     for (i in seq (objs))
     {
         mb <- microbenchmark::microbenchmark (
-           dat <- sf::st_read ("export.osm", layer=objs [i], quiet=TRUE) , 
-           times=times)
+           dat <- sf::st_read ("export.osm", layer = objs [i], quiet = TRUE),
+           times = times)
         size_sf <- c (size_sf, object.size (dat))
         mt_sf <- c (mt_sf, median (mb$time))
         cat ("\r", i, " / ", length (objs))
@@ -25,10 +25,10 @@ benchmark <- function (times=100)
     cat ("\rSF: Median times (in ms) for (", paste (objs), "):\n")
     cat ("\t(", mt_sf, "); total = ", sum (mt_sf), "\n")
 
-    mb <- microbenchmark::microbenchmark ( x <- osmdata_sf (q1, doc), 
-                                          times=times)
+    mb <- microbenchmark::microbenchmark ( x <- osmdata_sf (q1, doc),
+                                          times = times)
     #mb <- microbenchmark::microbenchmark ( x <- osmdata_sf (q1, "export.osm"),
-    #                                      times=10L)
+    #                                      times = 10L)
     mt <- median (mb$time / 1e6)
     cat ("osmdata: Median time = ", mt, " ms\n")
     size_od <- object.size (x)
