### 2017-03-26 #103 @mpadge
Okay, `gp()` no longer raises any issues other than unit tests and whole-package import, the latter  unavoidable for reasons described above. And only had to cheat with `#nolint` 4 times.
