### 2017-03-29 3268d49
add option to negate values
