### 2017-03-29 3e66d7a
update man/add_feature
