+++ R/opq.R
@@ -60,7 +60,7 @@ opq <- function (bbox=NULL)
 #'         add_feature("amenity", "pub") 
 #' c (osmdata_sf (q1), osmdata_sf (q1)) # all objects that are restaurants OR pubs
 #' # Use of negation to extract all non-primary highways
-#' q <- opq ("portsmouth uk") %>% add_feature (key="highway", value="!primary")
+#' q <- opq ("portsmouth uk") %>% add_feature (key="highway", value="!primary") 
 #' }
 add_feature <- function (opq, key, value, exact=TRUE, bbox=NULL)
 {
