+++ man/add_feature.Rd
@@ -41,7 +41,7 @@ q2 <- getbb ("portsmouth", display_name_contains="United States") \%>\% opq () \
         add_feature("amenity", "pub") 
 c (osmdata_sf (q1), osmdata_sf (q1)) # all objects that are restaurants OR pubs
 # Use of negation to extract all non-primary highways
-q <- opq ("portsmouth uk") \%>\% add_feature (key="highway", value="!primary")
+q <- opq ("portsmouth uk") \%>\% add_feature (key="highway", value="!primary") 
 }
 }
 \references{
