### 2017-03-29 2efcee4
impelement regex for add_feature
