### 2017-04-10 #103 @edzer
## Package Review of osmdata

It would be good to mention in the first sentence of the pkg description that it returns vector data, and not images (like pkg OpenStreetMaps does): not everyone knows what's behind the Overpass API.

##### Paper (for packages co-submitting to JOSS)

The package contains a `paper.md` with:

the target audience is not addressed in `paper.md`: is it end-users interested in exploring and plotting, or package developers using osmdata to ingest raw data?

`paper.md` does not give references to the packages used (imported) by osmdata, to packages or interfaces to which this package can be compared, or to the technology interfaced (open streetmaps, overpass api).

I didn't check for automated tests, or the guidelines.

Estimated hours spent reviewing: 6-8

---

### Design

This package imports openstreetmap vector data in R. It has functionality similar to osmar, which has not seen any development since 2013; osmar can export to sp, osmdata to sp and sf.  osmar, on the other hand, can do shortest paths through a network, and exposes the street network to R users (combining sp and igraph), osmdata does not expose the network. The package claims it is fast compared to sf/GDAL, but the link supporting this claim is dead.

The package advertises itself mostly mentioning the overpass api, but not everyone familiar with R and (vaguely with) openstreetmap will know what that means; it would be good to also mention that osmdata reads openstreetmap _vector_ data, as opposed to e.g. package OpenStreetMap which imports openstreetmap _raster_ data (images).

Some of the functions are rather verbose; e.g. function `osmdata` checks all data for missing, and if they are, assigns NULL. It would be much shorter to make NULL their default value; the function can also be written as
```
osmdata <- function (bbox = NULL, overpass_call = NULL, osm_points =
	NULL, osm_lines = NULL, osm_polygons = NULL, osm_multilines = NULL,
	osm_multipolygons = NULL, timestamp = NULL, ...) {

    structure(bbox = bbox, overpass_call = overpass_call, timestamp = timestamp,
        osm_points = osm_points, osm_lines = osm_lines, osm_polygons = osm_polygons,
        osm_multilines = osm_multilines, osm_multipolygons = osm_multipolygons,
		class = append(class(list()), "osmdata"))
}
```
which reduces 19 expressions to 1. Also, it is rather un R-like to make a class `c("list", "osmdata")`, instead of `c("osmdata", "list")` or even `c("osmdata"), since `list` is the default S3 superclass, and that would allow to add methods for `osmdata`, and fall back to `list` methods where useful. The `...` is not used (and `osmdata` is not a generic), and should be omitted, so that mis-typing of arguments lead to an error.

Function `osmdata` is supposed to be called internally only (not directly by users), and all the calls I could see look like `osmdata ()` (why the space before `()`?), so the whole list of arguments looks obsolete. The help page of `osmdata` gives an introduction to the whole package, to only reveal at the end that `osmdata()` should not be called. It mentions `osm class def`, but uses S3, where classes do not have a definition; somewhat confusing to those living in S3. I'm also not so sure whether the constructor function is needed at all: the class members are overwritten later without checking on using the $ operator; for a list L, a member that was not overwritten, like `L$whatever`, results in `NULL` also when it does not exist:
```r
> x = list()
> x$bbox
# NULL
```
this implies that `structure(list(), class = "osmdata")` could replace calls to `osmdata()`, allowing for removal of `osmdata()` altogether.

The functions starting with `osmdata_` export an overpass query to sp, sf or (write to file as) xml. It's somewhat confusing that
```
osmdata_sf(query, doc)
```
writes from doc into an object of class `sf`, where
```
osmdata_xml(query, doc)
```
writes the query to doc. Also, I had expected that option `quiet = TRUE` would suppress the message
```
q missing: osmdata object will not include query
```

The return of `osmdata_sp`, following the example, gives
```r
> hampi_sp
Object of class 'osmdata' with:
                 $bbox : 15.3159769,76.4406199,15.3559769,76.4806199
        $overpass_call : The call submitted to the overpass API
            $timestamp : [ Sun 1 Mar 2017 20:04:23 ]
           $osm_points : 'sp' SpatialpointsDataFrame with 109 points
            $osm_lines : 'sp' SpatiallinesDataFrame with 0 lines
         $osm_polygons : 'sp' SpatialpolygonsDataFrame with 18 polygons
       $osm_multilines : 'sp' SpatialmultilinesDataFrame with 0 multilines
    $osm_multipolygons : 'sp' SpatialmultipolygonsDataFrame with 0 multipolygons
```
where it might be better to use the class names of sp classes (SpatialpointsDataFrame etc. are not class names of sp, or of something else). Package sp does not distinguish between osm_lines and osm_multilines, or osm_polygons and osm_multipolygon; the list above confusingly suggests this. Why this distinction?

### Function names

Now, `osmdata_sf`, `osmdata_sp` and `osmdata_xml` have a large overlap; `osmdata_sf` and `osmdata_sp` could call `osmdata_xml` without the `file` argument to avoid this.

The vignette/README mention that
```
Or data can be read from a previously downloaded file:

x <- osmdata_sf(q1, "data.xml")
```
... but what is `q1` used for in this call? It seems that `q1` is added to the returned object, which takes its data from `data.xml`, but there is no guarantee that the two belong together. When the query call is added to the returned object (e.g. by `osmdata_xml`) this guarantee can be given, and ambiguity is excluded.

For the function names, it would help to memorize what they do if they would be, or contain, verbs; now, only `getbb` has a verb.

`osmdata_sf` is described as "Return an OSM Overpass query as an ‘osmdata’ object in ‘sf’ format."  I would have opted for "convert the result of an OSM Overpass query to a list with ‘sf’ objects".

### Install

During install, I see:
```
./configure: 5: [: Linux: unexpected operator
```
which points to a bug. The `configure` script aims at doing a coverage test on travis, but by checking for Linux. It might be better (esp. for linux users!) if it would test whether `USER` equals `travis` to create a conditional configuration for travis.

### Performance, comparison

Since the the paper mentions data is "processed with very fast C++ routines for return to R", I compared the speed to that of the GDAL/OGR driver, using `sf::st_read`:

```r
library(osmdata)
library(sf)

bb = c(11.561541, 48.14102, 11.582852, 48.1530)
q = opq(bbox = bb)
osmdata_xml(q, "x.osm")
system.time(o <- osmdata_sf(doc = "x.osm"))
layers = st_layers("x.osm")$name

osm_rd = function() { osmdata_sf(doc = "x.osm", quiet = TRUE) }
sf_rd = function() { system.time(lst <- lapply(layers, function(x) st_read("x.osm", x, quiet = TRUE))) }

library(microbenchmark)
microbenchmark::microbenchmark(osm_rd(), sf_rd(), times = 20)
# Unit: seconds
#      expr      min       lq     mean   median       uq      max neval
#  osm_rd() 2.393081 2.567723 2.645534 2.620662 2.708159 2.961327    20
#   sf_rd() 2.948154 3.195789 3.351276 3.318939 3.479391 3.760206    20

```
and see that the results are quite similar. Of interest is the difference between the objects returned:

```r
o
# Object of class 'osmdata' with:
#                  $bbox : 11.3675909042358 48.0759544372559 11.7228775024414 48.2664070129395
#         $overpass_call : The call submitted to the overpass API
#             $timestamp : [ Sun 1 Mar 2017 20:02:02 ]
#            $osm_points : 'sf' Simple Features Collection with 37997 points
#             $osm_lines : 'sf' Simple Features Collection with 4251 linestrings
#          $osm_polygons : 'sf' Simple Features Collection with 2682 polygons
#        $osm_multilines : 'sf' Simple Features Collection with 248 multilinestrings
#     $osm_multipolygons : 'sf' Simple Features Collection with 65 multipolygons

lst <- lapply(st_layers("x.osm")$name, function(x) st_read("x.osm", x, quiet = TRUE))
structure(sapply(lst, nrow), names = st_layers("x.osm")$name)
#          points            lines multilinestrings    multipolygons
#            6445             4447               52             2427
# other_relations
#             192
```
since both functions read the same xml file, I was suprised to see that the number of features returned differs pretty much. I then looked a bit further into the [GDAL OSM driver](http://www.gdal.org/drv_osm.html), which turns out to be highly configurable. Changing a few fields in the config file I got
```r
#          points            lines multilinestrings    multipolygons
#           37997             4569               52             2427
# other_relations
#             192
```
Which at least gets the same number of points, and even more lines and polygons, than osmdata. The number of options did scare me somewhat, and made me wonder which osmdata.ini configuruation would correspond to the osmdata query result, and how these configuration options are handled by osmdata. By the query call? If so, do R users need to learn the OSM Overpass API query call language?

### Minor things

The link to sfa (vignette 2) is http://www.opengeospatial.org/standards/sfa, not http://www.opengeospatial.org/standards/sfo

At several points, the vignettes misses the point that the word _simple_ in _simple features_ refers to "2D geometry with linear interpolation between vertices": simple (straight), as opposed to curved.

### Concluding

This is a nice package that shows a significant programming effort to get OpenStreetMap vector data into R, using the R datastructures of either sp or sf. It does not give much directions what users can, or should do with these data. Compared to package `osmar`, it is not clear whether (nor how) the network topology is obtained from an osmdata query result, how e.g.  the results from a query can be used to solve a shortest route problem in R.

The package claims to be considerably faster and to get more data, compared to the GDAL driver that is interfaced through e.g. R packages rgdal and sf.  I could confirm neither of these claims: reading through `sf::st_read` had a very similar speed, and the GDAL driver is highly configurable, allowing it to also read e.g. all the points that are vertices in lines and polygons. It would be good if osmdata could document to which GDAL driver options its default query corresponds: OSM API queries seem to be extremely configurable.

I wonder, for an end-users, what the value is of getting a point set that consists of not only point features (such as traffic lights, or way signs) but also all vertices in lines and polygons. The vignettes of osmdata do not point out how, from an osmdata query result, one could for instance isolate the point features for plotting purposes, nor other user benefits of having the ID of every possible point available.
