### 2017-04-12 b0d304c
minor tweak in response to edzers review
