+++ vignettes/osmdata.Rmd
@@ -526,7 +526,8 @@ all(is.na(addr))
 and `key=address` contains no data yet is still returned by `GDAL/sf`.
 
 Finally, note that `osmdata` will generally extract OSM data considerably faster
-than equivalent `sf/GDAL` routines (as detailed [here](git-wiki)).
+than equivalent `sf/GDAL` routines (as detailed
+[here](https://github.com/osmdatar/osmdata/wiki/Timing-benchmarks)).
 
 ### 3.3. The `osmdata_sp()` function
 
