### 2017-04-13 ropensci/osmdata#59 @mpadge
clarify distinction with OpenStreetMap package
This should be done, notably in response to @edzer's [ropensci review of the package](https://github.com/ropensci/onboarding/issues/103). We just need to work out the best place / places to do it. cc @Robinlovelace @maelle
