### 2017-04-13 osmdatar/osmdata#59 @maelle
README + intro vignette?
