### 2017-04-13 ropensci/osmdata#60 @mpadge
Function to extract only unique objects of each geom class
In response to @edzer's [ropensci review of the package](https://github.com/ropensci/onboarding/issues/103), it's be useful to implement a function to enable the standard `osmdata` output, which simply contains the entire data of all geometric types, to only those components of each geometric class that are not present in other classes. The result of this function would then be more comparable with the output of GDAL. cc @Robinlovelace @maelle
