### 2017-04-13 osmdatar/osmdata#59 @Robinlovelace
Yes to @maelle's suggestions - any easy fix.
