### 2017-04-13 osmdatar/osmdata#59 @mpadge
All agreed then. I'll happily volunteer for [the harder issue](https://github.com/osmdatar/osmdata/issues/60), so it'd be great if one of you could take care of that
