### 2017-04-13 osmdatar/osmdata#59 @mpadge
no rush: i won't be able to anything more until at least the week after anyway. plus it seems we've still got another two reviews coming.
