### 2017-05-17 #103 @mpadge
@ramnathv note that tests currently fail because of [this issue](https://github.com/edzer/sfr/issues/348) with the current CRAN version of `sf`. It's already fixed with `sf 0.5.0`, but tests won't pass again until that's on CRAN.
