### 2017-05-18 c9ab9c5
temp fix tests for sf 0.4.3; update README status wip -> active
