+++ tests/testthat/test-sf-construction.R
@@ -83,7 +83,8 @@ test_that ("sfc-point", {
 test_that ("sf-point", {
                x <- sf::st_sfc (sf::st_point(1:2))
                expect_identical (x, make_sfc (1:2, "POINT"))
-               y <- sf::st_sf (x)
+               #y <- sf::st_sf (x) # TODO: Revert for sf 0.5.0
+               y <- sf::st_sf (list ('x' = x)) # workaround for sf 0.4.3
                x <- make_sf (x)
                expect_identical (x, y)
 })
@@ -135,7 +136,8 @@ test_that ("sfc-line", {
 
 test_that ("sf-line", {
                x <- make_sfc (cbind (1:4, 5:8), "LINESTRING")
-               y <- sf::st_sf (x)
+               #y <- sf::st_sf (x) # TODO: Revert for sf 0.5.0
+               y <- sf::st_sf (list ('x' = x)) # workaround for sf 0.4.3
                x <- make_sf (x)
                expect_identical (x, y)
 })
@@ -159,7 +161,8 @@ test_that ("sf-multiple-lines", {
                x1 <- cbind (1:4, 5:8)
                x2 <- cbind (11:13, 25:27)
                x <- make_sfc (list (x1, x2), type = "LINESTRING")
-               y <- sf::st_sf (x)
+               #y <- sf::st_sf (x) # TODO: Revert for sf 0.5.0
+               y <- sf::st_sf (list ('x' = x)) # workaround for sf 0.4.3
                x <- make_sf (x)
                expect_identical (x, y)
 })
@@ -202,7 +205,8 @@ test_that ("sfc-polygon", {
 test_that ("sf-polygon", {
                xy <- cbind (c (1:4, 1), c(5:8, 5))
                x <- make_sfc (xy, "POLYGON")
-               y <- sf::st_sf (x)
+               #y <- sf::st_sf (x) # TODO: Revert for sf 0.5.0
+               y <- sf::st_sf (list ('x' = x)) # workaround for sf 0.4.3
                x <- make_sf (x)
                expect_identical (x, y)
 })
@@ -228,7 +232,8 @@ test_that ("sf-multiple-polygons", {
                xy1 <- cbind (c (1:4, 1), c(5:8, 5))
                xy2 <- cbind (c (11:14, 11), c(15:18, 15))
                x <- make_sfc (list (xy1, xy2), type = "POLYGON")
-               y <- sf::st_sf (x)
+               #y <- sf::st_sf (x) # TODO: Revert for sf 0.5.0
+               y <- sf::st_sf (list ('x' = x)) # workaround for sf 0.4.3
                x <- make_sf (x)
                expect_identical (x, y)
 })
@@ -274,7 +279,8 @@ test_that ("sfc-multilinestring", {
 test_that ("sf-multilinestring", {
                x <- make_sfc (cbind (c (1:4, 1), c (5:8, 5)),
                               type = "MULTILINESTRING")
-               y <- sf::st_sf (x)
+               #y <- sf::st_sf (x) # TODO: Revert for sf 0.5.0
+               y <- sf::st_sf (list ('x' = x)) # workaround for sf 0.4.3
                x <- make_sf (x)
                expect_identical (x, y)
 })
@@ -301,7 +307,8 @@ test_that ("sf-multiple-multilinestring", {
                x1 <- cbind (c (1:4, 1), c (5:8, 5))
                x2 <- cbind (c (11:13, 11), c (25:27, 25))
                x <- make_sfc (list (x1, x2), type = "MULTILINESTRING")
-               y <- sf::st_sf (x)
+               #y <- sf::st_sf (x) # TODO: Revert for sf 0.5.0
[…]
