### 2017-05-18 799f200
fix urls in vignettes; update license
