### 2017-05-18 353f3d4
add Rcpp registerRoutines + cran-comments.md; switch off one example
