+++ DESCRIPTION
@@ -14,9 +14,9 @@ Authors@R: c(
         comment="Author of included stub.R code"))
 Maintainer: Mark Padgham <mark.padgham@email.com>
 Title: Download OpenStreetMap Data with the Overpass API
-Description: Extraction of OpenStreetMap (OSM) data as 'sf' or 'sp' objects.
-    OSM data are extracted from the overpass API and processed with very fast
-    C++ routines for return to R.
+Description: Extraction of "OpenStreetMap" ('OSM') data as 'sf' or 'sp' objects.
+    'OSM' data are extracted from the 'Overpass' web server and processed with
+    very fast 'C++' routines for return to 'R'.
 Depends:
     R (>= 3.2.4)
 License: GPL-3 + file LICENSE
