+++ DESCRIPTION
@@ -19,7 +19,7 @@ Description: Extraction of "OpenStreetMap" ('OSM') data as 'sf' or 'sp' objects.
     very fast 'C++' routines for return to 'R'.
 Depends:
     R (>= 3.2.4)
-License: GPL-3 + file LICENSE
+License: GPL-3
 SystemRequirements: C++11
 NeedsCompilation: yes
 LazyData: true
