### 2017-05-18 6def08f
switch off all examples and all vignette code
