+++ vignettes/osmdata.Rmd
@@ -63,9 +63,12 @@ q <- opq(bbox = c(51.1, 0.1, 51.2, 0.2))
 Bounding boxes may also be defined by simply passing the name of a desired area,
 which does not have be precise:
 
-```{r opq2}
+```{r opq2, eval = FALSE}
 q <- opq(bbox = 'Greater London, U.K.')
 identical(opq(bbox = 'Greater London, U.K.'), opq(bbox = 'greater london uk'))
+```
+```{r echo = FALSE}
+TRUE
 ```
 
 Following the initial `opq()` call, `osmdata` queries are built by adding one or
@@ -73,8 +76,12 @@ more 'features', which are specified in terms of `key-value` pairs.  For
 example, all paths, ways, and roads are designated in OSM with `key=highway`, so
 that a query all motorways in greater London (UK) can be constructed as follows:
 
-```{r opq3}
+```{r opq3, eval = FALSE}
 q <- opq(bbox = 'greater london uk')
+q <- add_feature(q, key = 'highway', value = 'motorway')
+```
+```{r, echo = FALSE}
+q <- opq (bbox = c (51.2867602, -0.510375, 51.6918741, 0.3340155))
 q <- add_feature(q, key = 'highway', value = 'motorway')
 ```
 
@@ -84,11 +91,11 @@ A detailed description of features is provided at the
 comprehensive list of feature keys currently used in OSM.
 
 ```{r features, eval=FALSE}
-head(available_features())
+head (available_features ())
 ```
 
 ```{r, echo=FALSE}
-c("4wd only",  "abandoned", "abutters",  "access",    "addr",      "addr:city")
+c ("4wd only", "abandoned", "abutters", "access", "addr", "addr:city")
 ```
 
 There are two primary `osmdata` functions for obtaining data from a query:
@@ -127,10 +134,10 @@ as well as the
 
 `osmdata` queries are lists of class `overpass_query`. The actual query passed
 to the `overpass API` with a query can be obtained with the function
-`qry_to_string()`.  Applied to the preceding query, this function gives:
+`opq_to_string()`.  Applied to the preceding query, this function gives:
 
 ```{r, eval=FALSE}
-qry_to_string(q)
+opq_to_string(q)
 ## [out:xml][timeout:25];
 ## (
 ##   node
@@ -148,7 +155,7 @@ qry_to_string(q)
 
 The resultant output may be pasted directly into the 
 [overpass-turbo](http://overpass-turbo.eu/) online interactive query builder.
-(The output of `qry_to_string` has been somewhat reformatted here to reflect
+(The output of `opq_to_string` has been somewhat reformatted here to reflect
 the format typically used in `overpass-turbo`.)
 
 ### 2.1. osmdata queries
@@ -157,8 +164,13 @@ As demonstrated above, an `osmdata` query begins by specifying a bounding box
 with the function `opq()`, followed by specifying desired OSM features with
 `add_feature()`.  
 
-```{r kunming}
+```{r kunming, eval = FALSE}
 q <- opq(bbox = 'Kunming, China')
+q1 <- add_feature(q, key = 'natural', value = 'water')
+q2 <- add_feature(q1, key = 'name:en', value = 'Dian', value_exact = FALSE)
+```
+```{r, echo = FALSE}
+q <- opq(bbox = c(24.8915153, 102.5417638, 25.2115153, 102.8617638))
 q1 <- add_feature(q, key = 'natural', value = 'water')
 q2 <- add_feature(q1, key = 'name:en', value = 'Dian', value_exact = FALSE)
 ```
@@ -185,23 +197,23 @@ each feature added with `add_feature()`.
 The primary `osmdata` functions `osmdata_sf()` or `osmdata_sp()` pass these
 queries to `overpass` and return OSM data in corresponding `sf` or `sp` format,
[…]
