### 2017-05-18 fe015b1
stub out tests on travis as well as cran
