### 2017-05-18 2fed78d
Split-out query url
Ease debugging
