### 2017-05-18 897b36d
Split-out query url
Ease debugging
