### 2017-05-18 5e03546
Update handling of nominatim request - see #61
