### 2017-05-18 4d0ffc8
Fix #61 - seems to work
