### 2017-05-18 9995f56
switch off nominatim calls in tests
