+++ tests/testthat/test-getbb.R
@@ -45,11 +45,15 @@ test_that ('getbb-place_name', {
                        load("../cfm_output_bb.rda")
                        stub (getbb, 'httr::GET', function (x) cfm_output_bb )
                    }
-                   res <- getbb (place_name = "Salzburg")
-                   expect_is (res, "matrix")
-                   expect_length (res, 4)
-                   res <- getbb (place_name = "Salzburg", format_out = "string")
-                   expect_is (res, "character")
+                   # switching these tests off for now:
+                   if (!(is_cran | is_travis))
+                   {
+                       res <- getbb (place_name = "Salzburg")
+                       expect_is (res, "matrix")
+                       expect_length (res, 4)
+                       res <- getbb (place_name = "Salzburg", format_out = "string")
+                       expect_is (res, "character")
+                   }
                }
           })
 
