+++ DESCRIPTION
@@ -13,10 +13,10 @@ Authors@R: c(
     person ("Bartnik", "Lukasz", role=c("ctb","cph"),
         comment="Author of included stub.R code"))
 Maintainer: Mark Padgham <mark.padgham@email.com>
-Title: Download OpenStreetMap Data with the Overpass API
-Description: Extraction of "OpenStreetMap" ('OSM') data as 'sf' or 'sp' objects.
-    'OSM' data are extracted from the 'Overpass' web server and processed with
-    very fast 'C++' routines for return to 'R'.
+Title: Import OpenStreetMap Data as Simple Features or Spatial Objects
+Description: Download and import of "OpenStreetMap" ('OSM') data as 'sf' or 'sp'
+    objects.  'OSM' data are extracted from the 'Overpass' web server and
+    processed with very fast 'C++' routines for return to 'R'.
 Depends:
     R (>= 3.2.4)
 License: GPL-3
