### 2017-05-18 1e30558
update NEWS to CRAN release
