+++ NEWS.md
@@ -1,3 +1,3 @@
-0.0.0 
+0.0.0 (18 May 2017)
 ===================
-- Initial release
+- Initial CRAN release
