### 2017-05-19 805034f
rm configure and Makevars from R build; fix tests
