+++ NEWS.md
@@ -1,3 +1,8 @@
+0.0.1 (19 May 2017)
+===================
+- Remove configure and Makevars files
+- Fix tests
+
 0.0.0 (18 May 2017)
 ===================
 - Initial CRAN release
