### 2017-05-19 b59ab2d
fix tests so they all work offline
