+++ DESCRIPTION
@@ -37,7 +37,6 @@ Imports:
 Suggests:
     devtools,
     knitr,
-    microbenchmark,
     rmarkdown,
     roxygen2,
     sf,
