### 2017-05-21 76a7184
fix tests; switch on travis osx
