+++ tests/testthat/test-sf-osm.R
@@ -24,22 +24,19 @@ test_that ("multipolygon", {
                        dimnames (x$geometry [[1]] [[i]] [[j]]) <- NULL
                }
                names (x$geometry) <- NULL
-               if (!is_cran)
-                   expect_identical (x, x_sf)
-               else
-               {
-                   # CRS check fails on some R-devel windows machines
-                   expect_identical (names (x), names (x_sf))
-                   expect_identical (x$osm_id, x_sf$osm_id)
-                   expect_identical (x$name, x_sf$name)
-                   expect_identical (x$type, x_sf$type)
-                   g <- x$geometry
-                   g_sf <- x_sf$geometry
-                   attrs <- names (attributes (g))
-                   attrs <- attrs [attrs != 'crs']
-                   for (a in attrs)
-                       expect_identical (attr (g, a), attr (g_sf, a))
-               }
+               # CRS check fails on some R-devel windows machines
+               # expect_identical (x, x_sf)
+               expect_identical (names (x), names (x_sf))
+               expect_identical (x$osm_id, x_sf$osm_id)
+               expect_identical (x$name, x_sf$name)
+               expect_identical (x$type, x_sf$type)
+               g <- x$geometry
+               g_sf <- x_sf$geometry
+               attrs <- names (attributes (g))
+               #if (is_cran) # the cran machines do this check for some reason
+               attrs <- attrs [attrs != 'crs']
+               for (a in attrs)
+                   expect_identical (attr (g, a), attr (g_sf, a))
 })
 
 
@@ -58,22 +55,19 @@ test_that ("multilinestring", {
                    dimnames (x$geometry [[1]] [[i]]) <- NULL
                }
                names (x$geometry) <- NULL
-               if (!is_cran)
-                   expect_identical (x, x_sf)
-               else
-               {
-                   # CRS check fails on some R-devel windows machines
-                   expect_identical (names (x), names (x_sf))
-                   expect_identical (x$osm_id, x_sf$osm_id)
-                   expect_identical (x$name, x_sf$name)
-                   expect_identical (x$type, x_sf$type)
-                   g <- x$geometry
-                   g_sf <- x_sf$geometry
-                   attrs <- names (attributes (g))
-                   attrs <- attrs [attrs != 'crs']
-                   for (a in attrs)
-                       expect_identical (attr (g, a), attr (g_sf, a))
-               }
+               # CRS check fails on some R-devel windows machines
+               # expect_identical (x, x_sf)
+               expect_identical (names (x), names (x_sf))
+               expect_identical (x$osm_id, x_sf$osm_id)
+               expect_identical (x$name, x_sf$name)
+               expect_identical (x$type, x_sf$type)
+               g <- x$geometry
+               g_sf <- x_sf$geometry
+               attrs <- names (attributes (g))
+               #if (is_cran) # the cran machines do this check for some reason
+               attrs <- attrs [attrs != 'crs']
+               for (a in attrs)
+                   expect_identical (attr (g, a), attr (g_sf, a))
 })
 
 test_that ("ways", {
@@ -94,20 +88,17 @@ test_that ("ways", {
                # reset here
                attributes (x) <- attributes (x) [match (attributes (x_sf),
                                                         attributes (x))]
-               if (!is_cran)
-                   expect_identical (x, x_sf)
-               else
-               {
-                   # CRS check fails on some R-devel windows machines
-                   expect_identical (names (x), names (x_sf))
-                   expect_identical (x$osm_id, x_sf$osm_id)
-                   expect_identical (x$name, x_sf$name)
-                   expect_identical (x$type, x_sf$type)
-                   g <- x$geometry
-                   g_sf <- x_sf$geometry
[…]
