+++ .travis.yml
@@ -16,10 +16,14 @@ matrix:
         - os: linux
           r: devel
           dist: trusty
-          #- os: osx
-          #r: release
-          #before_install: ''
-          #disable_homebrew: true
+        - os: osx
+          r: devel
+          before_install: ''
+          disable_homebrew: true
+        - os: osx
+          r: release
+          before_install: ''
+          disable_homebrew: true
 
 before_install:
     - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
