### 2017-05-21 e68d243
vers bump; tidy .Rbuildignore; switch off OSX r-devel on travis
