+++ .travis.yml
@@ -20,10 +20,10 @@ matrix:
           r: devel
           before_install: ''
           disable_homebrew: true
-        - os: osx
-          r: release
-          before_install: ''
-          disable_homebrew: true
+          #- os: osx
+          #r: release
+          #before_install: ''
+          #disable_homebrew: true
 
 before_install:
     - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
