### 2017-05-21 50f6e75
fix travis.yml (osx r-devel -> osx r-release)
