### 2017-05-21 a99f0a0
add cran downloads to readme
