### 2017-05-22 0268264
fix memleak in trace-osm.cpp
