### 2017-05-22 85a0cb5
rm configure
