+++ src/Makevars.in
@@ -1,2 +1,0 @@
-PKG_CPPFLAGS=@cflags@
-PKG_LIBS=@libs@
