### 2017-05-23 #103 @karthik
Hi @mpadge, even though we did not receive the second review, we are pleased with Edzer's and also note that the package has been successfully submitted to CRAN. So I am pleased to accept this package into the rOpenSci suite! 🎉

I've created a new team called `osmdata` with you and me as members. Please go ahead and request a transfer of the pacakge the rOpenSci github account.

Once you transfer, please add the rOpenSci footer badge at the bottom of your README.

```
![](https://ropensci.org/public_images/github_footer.png)
```

Then you'll need to update CI links and badge links and also quickly reactivate CI services. We can help with this part.

Please reach out with any questions.
