### 2017-05-23 #103 @mpadge
Thanks @karthik - i've requested the transfer and just have to presume you get it (github notifications not so good there - i just got told i didn't have permission to do that). And thanks especially for boldly going ahead while still technically waiting for rev#2. I'll update CI guff as soon as transferred, and will also do the required DOI archiving. Two questions in that regard:
1. Will just one review be sufficient for JOSS, or will a second need to be requested?
2. Do or will I need to do anything further in regard to JOSS consideration, notably including informing them of DOI?
