### 2017-05-24 #103 @mpadge
@karthik All CI links and badges updated but I can't push becuase
```
# remote: Permission to ropensci/osmdata.git denied to mpadge.
```
