### 2017-05-24 #103 @karthik
> Will just one review be sufficient for JOSS, or will a second need to be requested?

Yes, that's sufficient. I'll handle your manuscript at the JOSS end.

> Do or will I need to do anything further in regard to JOSS consideration, notably including informing them of DOI?

Right before initiating the JOSS submission, let's make sure that the version you have submitted to CRAN is archived in Zenodo and has a DOI ready.

> # remote: Permission to ropensci/osmdata.git denied to mpadge.

bummer. Can you give me (@karthik) admin access on your original repo and I will quickly make the move and give you admin here?
