### 2017-05-31 #103 @mpadge
@karthik slight teething probs with CRAN - as soon as `0.0.2` is up, i'll archive it and let you know. (Submitted to CRAN over a week ago ...)
