### 2017-06-06 #103 @mpadge
@karthik finally deposited and doi'd [here](https://zenodo.org/record/803231). This version was submitted to CRAN on 22nd May, and should hopefully appear there sometime soon.
